refactor: Add improvements to UISearchBarController
The UISearchBarController now only sends and API request after the user has stopped typing a book or author title for 0.5 seconds. This will limit the amount of API requests that are made to the API server. Instead of an API request being made for every letter typed, one will be sent only when the user has finished typing for 0.5 seconds.